refactor: 로딩스피너 공용 컴포넌트로 이동
