+++ src/app/(with-nav-footer)/activities/[id]/ActivityDetail/index.tsx
@@ -13,7 +13,7 @@ import ReviewsSection from './_components/ActivityReviewsSection';
 import { MobileReservation, TabletReservation, DesktopReservation } from './_components/ActivityReservation';
 import ActivityBanner from './_components/ActivityBanner';
 import ScrollToTopButton from './_components/ScrollToTopButton';
-import LoadingSpinner from './_components/LoadingSpinner';
+import LoadingSpinner from '@/components/LoadingSpinner';
 
 const wrapper = 'mt-3 flex w-full flex-col gap-4 md:gap-6 max-w-[1200px]';
 const tabItems = [
